requirements.txt
pandas
requests
logzero
plotly
streamlit
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
xlsxwriter